refactor: move Talos client package to `pkg/`
As this implements Go client for Talos API, it makes sense to publish it
one the top level.
