+++ cmd/talosctl/cmd/talos/reboot.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client"
+	"github.com/talos-systems/talos/pkg/client"
 )
 
 // rebootCmd represents the reboot command
