+++ cmd/talosctl/cmd/talos/restart.go
@@ -12,7 +12,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/talos-systems/talos/api/common"
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client"
+	"github.com/talos-systems/talos/pkg/client"
 	"github.com/talos-systems/talos/pkg/constants"
 )
