+++ internal/integration/api/dmesg.go
@@ -13,8 +13,8 @@ import (
 	"time"
 
 	"github.com/talos-systems/talos/api/common"
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client"
 	"github.com/talos-systems/talos/internal/integration/base"
+	"github.com/talos-systems/talos/pkg/client"
 )
 
 // DmesgSuite verifies Dmesg API
