+++ internal/integration/api/logs.go
@@ -15,8 +15,8 @@ import (
 	"time"
 
 	"github.com/talos-systems/talos/api/common"
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client"
 	"github.com/talos-systems/talos/internal/integration/base"
+	"github.com/talos-systems/talos/pkg/client"
 	"github.com/talos-systems/talos/pkg/constants"
 )
