+++ internal/integration/api/reset.go
@@ -13,8 +13,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client"
 	"github.com/talos-systems/talos/internal/integration/base"
+	"github.com/talos-systems/talos/pkg/client"
 	"github.com/talos-systems/talos/pkg/retry"
 )
