+++ internal/pkg/provision/access/crashdump.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/talos-systems/talos/api/common"
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client"
+	"github.com/talos-systems/talos/pkg/client"
 	"github.com/talos-systems/talos/pkg/constants"
 )
