+++ internal/pkg/provision/check/service.go
@@ -9,8 +9,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client"
 	"github.com/talos-systems/talos/internal/pkg/provision"
+	"github.com/talos-systems/talos/pkg/client"
 	"github.com/talos-systems/talos/pkg/config/machine"
 )
