+++ pkg/config/types/v1alpha1/generate/talosconfig.go
@@ -7,7 +7,7 @@ package generate
 import (
 	"encoding/base64"
 
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client/config"
+	"github.com/talos-systems/talos/pkg/client/config"
 )
 
 // Talosconfig returns the talos admin Talos config.
