+++ pkg/config/types/v1alpha1/v1alpha1_configurator_bundle.go
@@ -5,8 +5,8 @@
 package v1alpha1
 
 import (
-	"github.com/talos-systems/talos/cmd/talosctl/pkg/client/config"
 	"github.com/talos-systems/talos/internal/pkg/runtime"
+	"github.com/talos-systems/talos/pkg/client/config"
 )
 
 // ConfigBundle defines the group of v1alpha1 config files.
